+++ src/edge/edge.h
@@ -16,15 +16,15 @@ struct edge {
 	unsigned int dest_index;
 	unsigned int x_dim;
 	unsigned int y_dim;
-	double joint_probabilities[MAX_STATES][MAX_STATES];
-	double message[MAX_STATES];
+	long double joint_probabilities[MAX_STATES][MAX_STATES];
+	long double message[MAX_STATES];
 };
 typedef struct edge* Edge_t;
 
-Edge_t create_edge(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, double **);
-void init_edge(Edge_t, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, double **);
+Edge_t create_edge(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, long double **);
+void init_edge(Edge_t, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, long double **);
 void destroy_edge(Edge_t);
-void send_message(Edge_t, double *);
+void send_message(Edge_t, long double *);
 
 
 #endif /* EDGE_H_ */
