+++ src/node/node.h
@@ -11,7 +11,7 @@
 #include "../constants.h"
 
 struct node {
-	double states[MAX_STATES];
+	long double states[MAX_STATES];
 	unsigned int num_variables;
 	unsigned int index;
 };
@@ -19,7 +19,7 @@ typedef struct node* Node_t;
 
 Node_t create_node(unsigned int index, unsigned int num_variables);
 void initialize_node(Node_t node, unsigned int index, unsigned int num_variables);
-void node_set_state(Node_t node, unsigned int num_variables, double * initial_state);
+void node_set_state(Node_t node, unsigned int num_variables, long double * initial_state);
 
 void destroy_node(Node_t);
